+++ lib/constants/library_list.dart
@@ -155,8 +155,15 @@ final List<Library> libraries = [
     license: 'GPL V3 License',
   ),
   Library(
-      name: 'pdf',
-      publisher: 'nfet.net',
-      url: 'https://github.com/DavBfr/dart_pdf',
-      license: 'Apache-2.0'),
+    name: 'pdf',
+    publisher: 'nfet.net',
+    url: 'https://github.com/DavBfr/dart_pdf',
+    license: 'Apache-2.0',
+  ),
+  Library(
+    name: 'receive_sharing_intent',
+    publisher: 'kasem.dev',
+    url: 'https://github.com/KasemJaffer/receive_sharing_intent',
+    license: 'Apache-2.0',
+  ),
 ];